refactor(app): split app logic into modules (#21)
* Refactor/tick elapsed (#20)

* refactor(app): centralize elapsed-time checks helper

* refactor(input): reuse shared elapsed-time helper

* refactor(app): split app loop into focused helper functions

* refactor(app): split app logic into domain modules
