+++ S-ADAPT/Core/Inc/input/input_utils.h
@@ -22,4 +22,9 @@ static inline uint8_t input_gpio_level(GPIO_TypeDef *port, uint16_t pin)
     return (HAL_GPIO_ReadPin(port, pin) == GPIO_PIN_SET) ? 1U : 0U;
 }
 
+static inline uint8_t input_has_elapsed_ms(uint32_t now_ms, uint32_t last_ms, uint32_t period_ms)
+{
+    return ((uint32_t)(now_ms - last_ms) >= period_ms) ? 1U : 0U;
+}
+
 #endif /* INPUT_UTILS_H */
